omnavning og omplassering så det går litt lettere an å følge bjørnlogikkken
