+++ src/Kart/LeafletTangram/scene/sceneFunksjoner/lagAktiveLag.js
@@ -1,0 +1,9 @@
+import lagEttLag from "./lagEttLag";
+
+function lagAktiveLag(aktive, viserKatalog, opplyst, config) {
+  Object.keys(aktive).forEach(kode =>
+    lagEttLag(aktive[kode], opplyst, viserKatalog, config)
+  );
+}
+
+export default lagAktiveLag;
