+++ src/Kart/LeafletTangram/scene/sceneFunksjoner/lagEttLag.js
@@ -1,0 +1,6 @@
+import byggLag from "./byggLag";
+
+export default function lagEttLag(lag, opplyst, viserKatalog, config) {
+  if (!lag.erSynlig && opplyst.kode !== lag.kode) return;
+  byggLag(lag, opplyst, config, viserKatalog);
+}
